Added tests and build environment
Added a vcpkg & CMake build environment.
vcpkg is used in manifest mode with a vcpkg.json file

Environment configuration files for Ubuntu 22.04 are included.
They should be run in the following order:
	sudo ./install_deps.sh
	source ./set_env.sh
	./vcpkg_install.sh

This will create a build directory.
Within the build directory, run:
	cmake ..
then:
	make

This will create a bin directory containing the binary.